pyproject.toml
[project]
name = "lawprocess"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "google-genai>=1.61.0",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
